requirements.txt
pyqrcode 
pypng